requirements.txt
apache-beam[gcp]==2.50.0
google-cloud-pubsub==2.18.4
google-cloud-bigquery==3.13.0
python-dotenv==1.0.0
